Extract Copilot session FS provider
Move in-memory Copilot session FS logic into a shared helper and reuse its config in CopilotStore. This deduplicates the session filesystem setup and keeps commit-message and conflict-resolution flows on the same provider implementation.
