+++ app/src/lib/stores/copilot-store.ts
@@ -5,8 +5,6 @@ import {
   AssistantMessageEvent,
   MessageOptions,
   SessionConfig,
-  SessionFsFileInfo,
-  SessionFsProvider,
 } from '@github/copilot-sdk'
 import { AccountsStore } from './accounts-store'
 import { Account, isDotComAccount } from '../../models/account'
@@ -34,6 +32,10 @@ import {
   IFileConflictContext,
   formatConflictContextForPrompt,
 } from '../copilot-conflict-context'
+import {
+  createCopilotInMemorySessionFsProvider,
+  getCopilotInMemorySessionFsConfig,
+} from '../copilot-in-memory-session-fs-provider'
 import * as ipcRenderer from '../ipc-renderer'
 import { startTimer } from '../../ui/lib/timing'
 import { join } from 'path'
@@ -46,7 +48,6 @@ import { enableCopilotSdkCommitMessageGeneration } from '../feature-flag'
 import type {
   Model,
   ModelBillingTokenPrices,
-  SessionFsReaddirWithTypesEntry,
 } from '@github/copilot-sdk/dist/generated/rpc'
 import { isGHE } from '../endpoint-capabilities'
 
@@ -157,231 +158,6 @@ export async function getCopilotCLIPath(): Promise<string> {
 
 function getCopilotCLIDir(): string {
   return join(__dirname, 'copilot')
-}
-
-interface IInMemorySessionFsFile {
-  readonly content: string
-  readonly createdAt: string
-  readonly updatedAt: string
-}
-
-function createInMemorySessionFsError(path: string): Error {
-  return Object.assign(new Error(`ENOENT: ${path}`), { code: 'ENOENT' })
-}
-
-function createInMemorySessionFsProvider(): SessionFsProvider {
-  const files = new Map<string, IInMemorySessionFsFile>()
-  const directories = new Set<string>(['.', 'state'])
-
-  const normalizePath = (path: string) => {
-    const normalized = path.replace(/\\/g, '/').replace(/\/+/g, '/')
-    return normalized === '' ? '.' : normalized.replace(/\/$/, '')
-  }
-
-  const getParentPath = (path: string) => {
-    const normalized = normalizePath(path)
-    const index = normalized.lastIndexOf('/')
-    return index > 0 ? normalized.slice(0, index) : '.'
-  }
-
-  const addDirectory = (path: string) => {
-    const normalized = normalizePath(path)
-    directories.add(normalized)
-
-    if (normalized !== '.') {
-      addDirectory(getParentPath(normalized))
-    }
-  }
-
-  const addParentDirectory = (path: string) => {
-    addDirectory(getParentPath(path))
-  }
-
-  const getTimestamp = () => new Date().toISOString()
-
-  const getDirectChildren = (path: string) => {
-    const normalized = normalizePath(path)
-    const prefix = normalized === '.' ? '' : `${normalized}/`
-    const children = new Set<string>()
-
-    for (const entry of [...files.keys(), ...directories]) {
-      if (entry === normalized || !entry.startsWith(prefix)) {
-        continue
-      }
-
-      const child = entry.slice(prefix.length).split('/')[0]
-      if (child.length > 0) {
-        children.add(child)
-      }
-    }
-
-    return [...children]
-  }
-
-  const writeFile = (path: string, content: string) => {
-    const normalized = normalizePath(path)
-    addParentDirectory(normalized)
-
-    const existing = files.get(normalized)
-    const timestamp = getTimestamp()
-    files.set(normalized, {
-      content,
-      createdAt: existing?.createdAt ?? timestamp,
-      updatedAt: timestamp,
-    })
-  }
-
-  return {
-    readFile: async path => {
-      const normalized = normalizePath(path)
-      const file = files.get(normalized)
-
-      if (file === undefined) {
-        throw createInMemorySessionFsError(path)
-      }
-
-      return file.content
-    },
-    writeFile: async (path, content) => writeFile(path, content),
-    appendFile: async (path, content) => {
-      const normalized = normalizePath(path)
-      const existing = files.get(normalized)
-      writeFile(normalized, `${existing?.content ?? ''}${content}`)
-    },
-    exists: async path => {
-      const normalized = normalizePath(path)
-      return files.has(normalized) || directories.has(normalized)
-    },
-    stat: async path => {
-      const normalized = normalizePath(path)
-      const file = files.get(normalized)
-
-      if (file !== undefined) {
-        const fileInfo: SessionFsFileInfo = {
-          isFile: true,
-          isDirectory: false,
-          size: Buffer.byteLength(file.content),
-          mtime: file.updatedAt,
-          birthtime: file.createdAt,
-        }
-        return fileInfo
-      }
-
-      if (directories.has(normalized)) {
-        const timestamp = getTimestamp()
-        const directoryInfo: SessionFsFileInfo = {
-          isFile: false,
-          isDirectory: true,
-          size: 0,
-          mtime: timestamp,
-          birthtime: timestamp,
-        }
-        return directoryInfo
-      }
-
-      throw createInMemorySessionFsError(path)
-    },
-    mkdir: async path => {
-      addDirectory(path)
-    },
-    readdir: async path => {
-      const normalized = normalizePath(path)
-      if (!directories.has(normalized)) {
-        throw createInMemorySessionFsError(path)
-      }
-
-      return getDirectChildren(normalized)
-    },
-    readdirWithTypes: async path => {
-      const normalized = normalizePath(path)
-      if (!directories.has(normalized)) {
-        throw createInMemorySessionFsError(path)
-      }
-
-      return getDirectChildren(normalized).map(name => {
-        const childPath = normalized === '.' ? name : `${normalized}/${name}`
-        const entry: SessionFsReaddirWithTypesEntry = {
-          name,
-          type: files.has(childPath) ? 'file' : 'directory',
-        }
-        return entry
-      })
-    },
-    rm: async (path, recursive, force) => {
-      const normalized = normalizePath(path)
-      const exists = files.has(normalized) || directories.has(normalized)
-
-      if (!exists) {
-        if (force) {
-          return
-        }
-
-        throw createInMemorySessionFsError(path)
-      }
-
-      if (directories.has(normalized)) {
-        const prefix = `${normalized}/`
-        const hasChildren = getDirectChildren(normalized).length > 0
-
-        if (hasChildren && !recursive) {
-          throw new Error(`Directory not empty: ${path}`)
-        }
-
-        for (const filePath of [...files.keys()]) {
-          if (filePath.startsWith(prefix)) {
-            files.delete(filePath)
-          }
-        }
-
-        for (const directoryPath of [...directories]) {
-          if (directoryPath.startsWith(prefix)) {
-            directories.delete(directoryPath)
-          }
-        }
-      }
-
-      files.delete(normalized)
-      directories.delete(normalized)
-    },
-    rename: async (src, dest) => {
-      const normalizedSrc = normalizePath(src)
-      const normalizedDest = normalizePath(dest)
-      const file = files.get(normalizedSrc)
-
-      if (file !== undefined) {
-        addParentDirectory(normalizedDest)
-        files.set(normalizedDest, file)
-        files.delete(normalizedSrc)
-        return
-      }
-
-      if (!directories.has(normalizedSrc)) {
-        throw createInMemorySessionFsError(src)
-      }
-
-      addDirectory(normalizedDest)
-
-      const srcPrefix = `${normalizedSrc}/`
-      const destPrefix = `${normalizedDest}/`
-      for (const [filePath, entry] of [...files]) {
-        if (filePath.startsWith(srcPrefix)) {
-          files.set(`${destPrefix}${filePath.slice(srcPrefix.length)}`, entry)
-          files.delete(filePath)
-        }
-      }
-
-      for (const directoryPath of [...directories]) {
-        if (directoryPath.startsWith(srcPrefix)) {
-          directories.add(
-            `${destPrefix}${directoryPath.slice(srcPrefix.length)}`
-          )
-          directories.delete(directoryPath)
-        }
-      }
-
-      directories.delete(normalizedSrc)
-    },
-  }
 }
 
 /**
@@ -990,11 +766,7 @@ export class CopilotStore extends BaseStore {
         }`,
       },
       workingDirectory: repositoryPath,
-      sessionFs: {
-        initialCwd: repositoryPath ?? process.cwd(),
-        sessionStatePath: 'state',
-        conventions: __WIN32__ ? 'windows' : 'posix',
-      },
+      sessionFs: getCopilotInMemorySessionFsConfig(repositoryPath),
       gitHubToken: account.token,
     })
   }
@@ -1236,7 +1008,7 @@ export class CopilotStore extends BaseStore {
           },
           availableTools: [],
           enableSessionStore: false,
-          createSessionFsProvider: createInMemorySessionFsProvider,
+          createSessionFsProvider: createCopilotInMemorySessionFsProvider,
           onPermissionRequest: async () => ({
             kind: 'reject',
           }),
@@ -1534,7 +1306,7 @@ export class CopilotStore extends BaseStore {
         streaming: true,
         availableTools: [],
         enableSessionStore: false,
-        createSessionFsProvider: createInMemorySessionFsProvider,
+        createSessionFsProvider: createCopilotInMemorySessionFsProvider,
         systemMessage: {
           mode: 'append',
           content: ConflictResolutionSystemPrompt,
